+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 This list is not intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate. Given version `A.B.C.D`, breaking changes are to be expected in version number increments where changes in the `A` or `B` sections:
 
+### v3.17.2.0
+- **http4k-connect-*** - Upgrade dependencies, including Kotlin to 1.6.21.
+- **http4k-connect-amazon-dynamodb-** - Add auto Lens extensions for object mapping: `Moshi.autoDynamoLens<AnObject>()`
+
 ### v3.17.1.0
 - **http4k-connect-*** - Upgrade dependencies, including Kotlin to 1.6.20.